# docs/03-Architecture/openapi.yaml
openapi: 3.1.0
info:
  title: ARENAX Core API
  version: 0.1.0
security:
  - cookieAuth: []
paths:
  /api/v1/auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginInput'}
      responses:
        '200':
          description: Authenticated; an HttpOnly access cookie is issued.
          content: {application/json: {schema: {$ref: '#/components/schemas/LoginResult'}}}
        '401': {description: Invalid credentials}
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      responses:
        '200': {description: Current User, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}
        '401': {description: Access absent, expired or revoked}
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses: {'204': {description: Access revoked and cookie removed}}
  /api/v1/people:
    get:
      operationId: listPeople
      responses:
        '200':
          description: People ordered by name.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/NamedResource'}}}}
    post:
      operationId: createPerson
      requestBody: {$ref: '#/components/requestBodies/NamedResource'}
      responses: {'201': {$ref: '#/components/responses/Created'}}
  /api/v1/spaces:
    get:
      operationId: listSpaces
      responses:
        '200':
          description: Spaces ordered by name. Occupancy is derived from Sessions and is not returned here.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Space'}}}}
    post:
      operationId: createSpace
      requestBody: {$ref: '#/components/requestBodies/NamedResource'}
      responses: {'201': {$ref: '#/components/responses/Created'}}
  /api/v1/equipments:
    get:
      operationId: listEquipments
      parameters:
        - {name: space_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Equipment ordered by kind and external identifier.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Equipment'}}}}
    post:
      operationId: createEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EquipmentInput'}
      responses: {'201': {$ref: '#/components/responses/Created'}}
  /api/v1/sessions:
    get:
      operationId: getAgenda
      summary: List Sessions whose scheduled period intersects the requested window.
      parameters:
        - {name: start, in: query, required: true, schema: {type: string, format: date-time}}
        - {name: end, in: query, required: true, schema: {type: string, format: date-time}}
        - {name: space_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Sessions ordered by scheduled start.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Session'}}
    post:
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SessionInput'}
      responses:
        '201': {$ref: '#/components/responses/Session'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/sessions/{sessionId}/actions/{action}:
    post:
      operationId: transitionSession
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: action, in: path, required: true, schema: {type: string, enum: [start, finish, cancel]}}
      responses: {'200': {$ref: '#/components/responses/Session'}, '409': {$ref: '#/components/responses/Conflict'}}
  /api/v1/sessions/{sessionId}:
    get:
      operationId: getSessionDossier
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Session dossier including Moments and Timeline}
        '404': {description: Session not found}
  /api/v1/sessions/{sessionId}/extend:
    post:
      operationId: extendSession
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduled_end]
              properties: {scheduled_end: {type: string, format: date-time}}
      responses: {'200': {$ref: '#/components/responses/Session'}, '409': {$ref: '#/components/responses/Conflict'}}
  /api/v1/events/button-pressed:
    post:
      security: []
      operationId: registerButtonPress
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ButtonPress'}
      responses:
        '202':
          description: Physical event audited; accepted indicates whether a Moment was requested.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ButtonPressResult'}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: arenax_acesso
  requestBodies:
    NamedResource:
      required: true
      content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string}}}}}
  responses:
    Created:
      description: Created
      content: {application/json: {schema: {type: object, properties: {id: {type: string, format: uuid}}}}}
    Session:
      description: Session state
      content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}
    Conflict:
      description: Domain invariant rejected the intention
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
        remember: {type: boolean, default: false}
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        email: {type: string, format: email}
        role: {type: string, enum: [proprietario, administrador, operador]}
    LoginResult:
      type: object
      required: [user]
      properties: {user: {$ref: '#/components/schemas/User'}}
    NamedResource:
      type: object
      required: [id, name]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
    Space:
      allOf:
        - {$ref: '#/components/schemas/NamedResource'}
        - type: object
          required: [administrative_status]
          properties: {administrative_status: {type: string}}
    Equipment:
      allOf:
        - {$ref: '#/components/schemas/EquipmentInput'}
        - type: object
          required: [id]
          properties: {id: {type: string, format: uuid}}
    EquipmentInput:
      type: object
      required: [space_id, kind, external_id]
      properties:
        space_id: {type: string, format: uuid}
        kind: {type: string, enum: [camera, ax_device]}
        external_id: {type: string}
        configuration: {type: object}
    SessionInput:
      type: object
      required: [responsible_person_id, space_ids, scheduled_start, scheduled_end]
      properties:
        responsible_person_id: {type: string, format: uuid}
        space_ids: {type: array, minItems: 1, items: {type: string, format: uuid}}
        scheduled_start: {type: string, format: date-time}
        scheduled_end: {type: string, format: date-time}
    Session:
      allOf:
        - {$ref: '#/components/schemas/SessionInput'}
        - type: object
          properties:
            id: {type: string, format: uuid}
            status: {type: string}
            actual_start: {type: [string, 'null'], format: date-time}
            actual_end: {type: [string, 'null'], format: date-time}
    ButtonPress:
      type: object
      required: [device_id, timestamp]
      properties: {device_id: {type: string}, timestamp: {type: string, format: date-time}}
    ButtonPressResult:
      type: object
      required: [accepted, reason]
      properties:
        accepted: {type: boolean}
        reason: {type: string}
        moment_id: {type: [string, 'null'], format: uuid}
        session_id: {type: [string, 'null'], format: uuid}
    Error:
      type: object
      properties: {code: {type: string}, message: {type: string}}